+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @authhero/kysely-adapter
 
+## 2.0.0
+
+### Minor Changes
+
+- Add table for refresh tokens
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.38.0
+
 ## 1.0.0
 
 ### Minor Changes